+++ SPOTeam_android/app/src/main/res/drawable/custom_rectangle_drawable.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="#D6E5FF" /> <!-- 사각형 배경 색상 -->
+    <corners android:radius="8dp" /> <!-- 모서리 반지름 -->
+
+    <size android:width="20dp" android:height="20dp" /> <!-- 사각형 크기 -->
+</shape>
