+++ SPOTeam_android/app/src/main/res/values/themes.xml
@@ -165,4 +165,9 @@
         <item name="android:textColor">@drawable/my_date_text_color</item>
     </style>
 
+    <style name="CustomDateTextAppearance" parent="TextAppearance.MaterialComponents.Body1">
+        <item name="android:textColor">#000000</item> <!-- 검정색 텍스트 -->
+        <item name="android:textSize">12sp</item> <!-- 텍스트 크기 -->
+    </style>
+
 </resources>
